neomutt: 20211015
turns out we can make the manpages without an internet connection.
this ./configures as if it would build the docs, so that the Makefile contains the logic
required to build the manpages, but then the all-docs target is removed from ALL_TARGETS.
We later call make separately just to build what we want.